+++ szukaj.h
@@ -2,7 +2,7 @@
 #ifndef SZUKAJ_H
 #define SZUKAJ_H
 
-int hash(char *, int, long int*); //funkcja hashujaca - oblicza "wartosc" podanego slowa wg algorytmu
+int hash(char *, int, unsigned long long int*); //funkcja hashujaca - oblicza "wartosc" podanego slowa wg algorytmu
 // bool check(char *); //pierwsze sprawdzenie, jesli nie przejdzie brak zgodnosci/ jesli zwroci true - podejrzenie o zgodnosc
 int check(char *, char *); //drugie sprawdzenie, malo efektywne ale wykonywane nie wiele razy, maly wplyw na zlozonosc obliczeniowa algorytmu
 int length(char *); //zwraca dlugosc ciagu znakow
